[UO266575] - Tests comentados y validación eliminar ofertas de otros …

+++ src/main/java/com/uniovi/sdimywallapop/controllers/OffersController.java
@@ -167,6 +167,10 @@ public class OffersController {
         String email = principal.getName();
         User user = usersService.getUserByEmail(email);
         List<Long> offers = offersService.getOffersIdsByUserId(user.getId());
+        List<String> errors = offersService.validateOfferToDelete(offersService.getOffer(id), user);
+        if (errors.size() > 0) {
+            return "redirect:/offer/myList";
+        }
 
         if(offers.contains(id)){
             offersService.deleteOffer(id);
